Add a 'shell' subcommand to ekiden-tools, per #80
* refactored docker calls. wait for docker to finish before exiting.
* container name includes hash of working dir
* switch bash scripts to ekiden-tools shell
* refactor subcommands to new mods
* allow skipping intel verification per #98

Use a compile time enviornment variable to disable AVR verification.

This should be a runtime configurable thing, but it's exceedingly
non-trivial to actually make this runtime configurable because the key
manager contract is instantiated via a global static constructor, among
other reasons.

WARNING: Contracts built on containers that pre-date this commit will not
work as the enviornment variable to disable AVR verification will not be
set.  To fix this either rebuild the Docker container or set
`EKIDEN_UNSAFE_SKIP_AVR_VERIFY=1` within the container.