Use a compile time enviornment variable to disable AVR verification.
This should be a runtime configurable thing, but it's exceedingly
non-trivial to actually make this runtime configurable because the key
manager contract is instantiated via a global static constructor, among
other reasons.

WARNING: Contracts built on containers that pre-date this commit will not
work as the enviornment variable to disable AVR verification will not be
set.  To fix this either rebuild the Docker container or set
`EKIDEN_UNSAFE_SKIP_AVR_VERIFY=1` within the container.