The start of ethereum stake contract.
Much of the infrastructure code is cargo-culted from @yawning's registry.
The Stake.sol is an approximation of the rust interface. In particular,
where the rust interface returned a struct or a Vec<>, solidity cannot, so
the API differs.

Should we change the Rust version and the API proto file to match? Or
just make the rust<->ethereum web3 adapter make it work?