compute, db, ethereum: lower storage expiry
fixes #409

Transient items are stored for one epoch or two if they'll be useful
to the next epoch's committee or if there's a concern about
availability during an epoch transition.

Registrations are stored for 90 epochs, which for an epoch interval of
one day, is a common validity range for TLS certificates.