basic testing of get_name.  fetch_escrow_by_id is plumbed in but untested.
plan to get returned value from a contract call from web3 is to use
call_with_confirmations always to get the block number where the
transaction got committed, and since as long as no byzantine actors
use call to pack more than one value-returning transaction into a
block, we can use query at the previous block to get the value that
the real call had returned.  We could also do query at the current
block first, before doing call_with_confirmations.  This is good
enough for version A.

For subsequent versions, we will have to be sure our assumptions are
in order.  If the Entity's or the compute node's keys were leaked or
act in a Byzantine manner, then it'd be easy for multiple transactions
to appear in a block (for example).  There is no hope of introducing a
nonce to identify the event from the transaction that we want the
results from versus from another one (a form of replay / message
re-use), since the request is public info and an adversary can snoop
that and win a race to get their transaction processed first.