Heap-based assignment of subgraphs to committees
We use a heap to order the committees based on completion time (sum of
estimated execution time for all transactions assigned to the
committee so far).  With batches (subgraphs) ordered in decreasing
work time, we delmin to find the first committee that will finish, and
assign the most expensive unassigned batch to it.  This is just the
first-worker-available heuristic, rather than trying to do a variation
of bin packing (NP-hard).  (Is this variation actually hard?)

We may want to look into MTP or Korf's algorithm if this heuristic is
not sufficiently good for our needs.  The assignment of batches to
committees is only in our simulator right now, but should eventually
be moved into the alg package proper.