improve verbose sim parameters output
Made each grouped parameter struct also include a Show() method so
that at non-zero verbosity levels, we can have a nice display of all
the configuration parameters.  This should mean that if this output is
saved, it will be easy to figure out how to set command-line
parameters to recreate that execution.