test Generate
The test is not particularly good, just that the output is in the
expected range.  Also modified to allow command-line flag to set the
rng seed and to output it (test output is suppressed unless -v is
given to the "go test" command, or if the test fails), so that if it
ever failed it'd be easy to reproduce.