CR: Fixed excess-fraction flag description, logger name, nolint
The description for the excess-fraction flag for proportional feedback
was poorly worded.  This improves it by explaining that the -adaptive
scheduler uses it, and how.

The logger name is now derived from code path. The nolint comment
placement is at the end-of-line for the first line of code for the
function which it affects