Add randomly sized conflict sets
This CL adds randomly sized conflict sets.  The
RandomDistributionRandNumLocationsTransactionSource (ugh) replaces
RandomDistributionTransactionSource to accept two Rngs instead of two
integers for the sizes of the read and write sets.  New Rngs includes
Fixed -- which provides the ability to have fixed values just like
pre-CL -- and Add which adds the outputs of Rngs, so that we can
easily have the write-set minimal value to be the 2 (and have its
probability be the zipf 1/normalizer, 3 associated with 1/(2^alpha
normalizer), etc).  This Cl also fixes a bug where we were too eager
to reset() an iterflag controlled flag, which causes problems with the
non-adaptive scheduler.