go/tendermint/db/bolt: Enable snappy compression
This blanket enables Snappy compression of all values stored in the
tendermint databases that are backed by BoltDB, as a quick check with
the e2e test indicates that all databases will benefit.

Note: This change is BACKWARD INCOMPATIBLE.  Migration code is omitted
as the immediate prior change requires a state wipe regardless,
rendering having such a thing moot.

Implements #1032.