go/worker: Add support for overriding registered P2P address
We already had support for overriding the registered address used for
client gRPC connections, but we need something similar for the P2P
address.

This commit adds a new option --worker.p2p.addresses which behaves the
same way as --worker.client.addresses but is used for P2P addresses.