go/tendermint: Fix the keying to be compatible with 0.30.1
 * Unify the tendermint (private) validator key and p2p link key.
 * Store the hex encoded public keys in the address book as lower case
   strings instead of upper case strings.

Technically we could probably keep the validator key and p2p link key
separate if we extended the gensis document format again, but this is
easier and will do till we revisit all of our keying and properly
carve out purpose-specific keys for everything.

And yes, the "fix" in 0.30.1 for keying is to add a case-sensitive
string comparision based public key check.  I'm not sure if I should
laugh or cry.