runtime: Use a persistent RAK, derived from EGETKEY
This is somewhat of an abuse of the SGX Sealing key, but it should be
safe to use it in this context (additionally there is sufficient domain
separation, both at the EGETKEY instruction level, and at the KDF
applied after).

Note: If we ever decide to enable RAKs for "simulation mode", all
instances will have the same key.