go/storage/badger: Add a Badger based backend
I'm not sure how I feel about this to be honest.  According to the
microbenchmarks, goleveldb is marginally faster for the single test
cases, but Badger is dramatically faster for the batch inserts.

That said, at least on my laptop, the microbenchmarks failed to run to
completion for Badger due to the OOM killer kicking in and killing the
processs.  For the sake of being thorough, I haven't tried it with the
Go GC behavior set to be more kernel stupidity friendly, nor did I try
it writing to disk as opposed to tmpfs.  But goleveldb runs fine even
without taking such extra steps.

It's probably ok to keep around as a experimental thing people can try,
since I refactored out a lot of the common code between implementations
in my previous commit.