go/genesis: Sign the genesis validators with an entity
I still have no idea what the best way to handle non-genesis validators
is.   The current plan is likely to add them to the registry as part of
the node-descriptor, but that will create interesting issues with the
dump/restore process.

Note: The bootstrap server will use the debug test entity to sign the
validator, so bad things will happen if it's not configured to use test
keys or the test entity.