go/tendermint/apps/roothash: Stop timer when blocks are emitted
Since we have two timeouts (for compute and merge committee), one timer
can still be left armed when a round fails. This makes sure that if a
block is emitted, the timer is disarmed.

Also makes spurious timeouts a hard error as those should never happen
and it is better to fail if it does.