misc: Go back to serializing certain things as CBOR arrays
The Go CBOR library we are using gained support for `toarray`, so revert
the various structures that used to be serialized that way to save space
when we were using go-codec.

Note: While the cbor library supports both `cbor:` and `json:` and we
favor `json:` to make life easier down the road, for structs that use
`toarray`, I opted for `cbor:` since the option is non-standard.