go tendermint staking: adjust total supply in runtime transfer
We previously left the total supply inconsistent when transferring
tokens to/from a runtime. Fix that to adjust the total supply
accordingly when transferring.

It was also proposed to move these transferred tokens to/from
per-runtime holding areas for bookkeeping. That remains a good
idea.

It's proposed in #2377 to dismantle this whole runtime transfer
system. So with any luck, none of this will be used after all.