go/oasis-node: Add the `unsafe-reset` sub-command
As it probably can be deduced from the name, this sub-command will
reset the node back to a freshly provisioned state (preserving any
key material if it exists).

The `-n` (`--dry_run`) flag will go through the process without actually
removing state, so that it is possible to preview the results of
executing the command.

Usage: `oasis-node unsafe-reset --datadir=/path/to/datadir`