runtime: Persist (sealed) RAK in untrusted local storage
Previously the RAK was totally ephemeral and got regenerated on each runtime
enclave restart.

This caused problems in case an enclave got restarted in the middle of an epoch
as the RAK only got updated during a node's re-registration (which is once per
epoch). Until the re-registration and an epoch transition happened, all
RAK-signed messages were rejected.