go/worker: Make worker registration event-based
Previously, registration was based exclusively on epoch transitions
which meant that in case the node was not yet ready to service requests
when an epoch transition happened, it needed to block the registration
goroutine until it was ready.

This prevented updating the registration in reaction to some worker
event that causes the node descriptor to become invalid (e.g., due to a
runtime restart invalidating the RAK).

This commit changes the worker registration as follows:

- Each part of the code that wishes to contribute something to the node
  descriptor first asks the registration worker to create a
  RoleProvider. This reserves a slot which is "not ready" by default.

- The registration worker is changed so that it reacts to both epoch
  transitions and role provider changes. When a change is detected, the
  registration worker will check if all RoleProvider slots indicate that
  the slot is "ready". In case any is not ready, registration will not
  proceed.

- Anyone that has a reference to a RoleProvider can signal that the slot
  has become either "ready" or "not ready".