go/staking: Add stateful stake accumulator
Previously there was no central place to account for all the parts that need to
claim some part of an entity's stake which required approximations all around.

This is now changed and a stateful stake accumulator is added to each escrow
account. The stake accumulator is used to track stake claims from different
apps (currently only the registry).

This change also means that node registration will now always require the
correct amount of stake.