storage/mkvs: Only nil value should mean deletion
Previously an empty value in the write log signalled that the given entry is a
delete operation instead of an insert one. This was incorrect as inserting an
empty byte string is allowed. This changes

This change is BREAKING as it changes write log semantics and thus it breaks
the runtime worker-host protocol.