storage: Rename "round" to "version"
Previously the MKVS used the term "round" to mean a monotonically increasing
version number. This choice was due to the fact that it was initially used to
only store runtime state which has a concept of rounds.

As we expand its use it makes more sense to generalize this and call it version.