keymanager-lib: Bind persisted state to the runtime ID
It is likely prudent to bind the persisted master secret to the runtime
ID.  This change does so by including the key manager runtime ID as the
AAD when sealing the master secret.

This is backward incompatible with all current key manager instances as
the existing persisted master secret will not decrypt.