rust: Remove the use of the `failure` crate
Using this crate is no longer considered best practice, and bugs in it
have broken the build before.

Note: Unfortunately the `runtime-loader` crate still uses this `failure`
nonsense since the external Intel AESM client crate uses it for error
handling.  Since it is a stand-alone binary it has been left as is.