go/worker/compute/merge: Drop support for multiple committees
Since there is currently no transaction scheduler implementation which would
support multiple committees, there is no sense in the merge node to try to
support such cases as it could be a source of bugs. Additionally it results
in extra round trips to storage nodes due to the Merge operation which in
case of a single committee does not do anything.

The merge node is also the only client for the Merge* storage operations, so
they can just be removed in order to reduce the exposed API surface.