go: Fix CLI tools to not print error messages twice
Previously, Execute() called nodeCommon.EarlyLogAndExit() if there was
an error running rootCmd.Execute() which in turn printed the error to
stderr.
The error message is printed to stdout via github.com/spf13/cobra's
ExecuteC() already, so this resulted in the same error message being
printed twice.

This is an expansion of f5f906b3c076290d20ea6cbd298ce548c2000513 which
fixed this behavior for go/oasis-test-runner/cmd.