Compare protocol versions according to SemVer 2.0.0
Since we bumped the protocol versions to version 1.0.0 with the release
of Oasis Core 20.10, we also need to modify how we compare and detect
backward-incompatible changes to follow SemVer 2.0.0 rules.

From now onwards, only a change in a protocol's major version signifies
a backward-incompatible change.

go/common/version: Introduce TendermintAppVersion variable which is the
Tendermint ABCI application's version and should be compatible with
Tendermint's version checks.