go/runtime/client: Add GetTransactions method, deprecate GetTxs
A new method called GetTransactions has been added to replace the now deprecated
GetTxs. The new method does not require the caller to pass the IO root as that
is an internal detail. This makes it consistent with the existing GetEvents API.

Users of GetTxs should migrate to GetTransactions as the former may be removed
in a future release.