go/runtime: Add key manager EnclaveRPC support to client runtime hosts
Previously runtimes hosted on client nodes were not able to issue EnclaveRPC
requests as the endpoint was not implemented in the handler. This unifies the
handler implementations between the client and compute nodes and makes sure that
key manager requests are possible (of course client nodes will only be able to
request public keys).