go/roothash: Refactor executor commitments
Previously commitments were constructed in such a way that proving
equivocation for proposers required inclusion of all emitted messages as
part of evidence. Given that the messages can be big this is far from
optimal and this has now been changed such that only the hashes need to
be included.