go/worker/storage: Remove separate storage sync status store
Previously the worker maintaned a separate store that kept information
about the progress of storage sync. Since it was a separate store this
could cause problems if it got out of sync (e.g. due to partial manual
copies). This should make the process more robust as there is only one
source of truth.