go/worker/registration: Clear deregister flag when deregistered
When gracefully halting the node after the node's registration
expires, to ensure that the deregistration and shutdown occurs,
the node will persist the fact that it is mid-shutdown in a flag.

Previously, this flag had to be cleared manually by the node operator
which, while serving to ensure that the node does not get restarted
and re-register, is sub-optimal as it required manual intervention.

Instead, if the node is deregistered cleanly, we will clear the flag
under the assumption that the operator can configure whatever
automation they are using to do the right thing.