go/consensus/tendermint/full: Unify started indicators
Previously there were two sets of "node started" indicators, one for the
commonNode and the other one for the fullService/archiveService. Due to
how things were initialized this could cause the full node to report
that it is "started" too soon which would cause some queries to trigger
a segmentation fault instead of blocking.