mkvs/proof: Make sure that all entries in the proof have been used
Note that this does not impact verification correctness as only visited
nodes are part of the proof computation and the verified tree returned
by the verifier.

But in case future code may want to do something with raw entries of a
verified proof this removes a possible footgun where unverified entries
could be considered.