go/p2p: Improve peer manager's peer handling
Peer manager currently connects to all nodes in the registry which has multiple
drawbacks (connections get pruned, no protocol selection, no peer discovery).
This should be changed so that peers are connected depending on the supported
protocols and the number of peers in a protocol/topic should be regulated
in a controlled fashion.