go/worker/storage: Fix case when checkpoint sync disabled but forced
If checkpoint sync is disabled but sync has been forced (e.g. because
the state at genesis is non-empty), we must request to sync the
checkpoint at genesis as otherwise we will jump to a later state which
may not be desired given that checkpoint sync has been explicitly
disabled via config.