runtime/consensus/tendermint/verifier: Correctly compare headers
Since the store may have an earlier (non-canonical, but valid) version
of the block available, we need to only compare the actual header and
not the commits/signatures.

This is because it can happen that during the immediate sync the light
block does not yet contain all of the commits (but only just enough to
be valid, e.g. 2/3+) and this gets stored in the light block store.
Later on (e.g. during a query) the presented light block may have the
full set of commits.