runtime/src/enclave_rpc: Support calls to explicit key manager members
Key manager enclaves can now request a host to talk not only to a randomly
chosen key manager instance, but also to a specific instance. The identity
of the remote node is verified only in Noise sessions. In these, the enclave
obtains the other instance's trusted RAK from the consensus layer and compares
it to the one used throughout the session. These changes also eliminate the
need for a sticky P2P key manager node, which could be removed in the future.