go/consensus/tendermint: Remove low-level requests from services
Previously the raw RequestBeginBlock request was distributed to the
various consensus services. This could result in these services using
invalid state (e.g. using the req.Hash in the proposal preparation phase
when this data is not yet available).

These requests are no longer propagated and any relevant state is made
available through the block context.