go/oasis-test-runner: Fix flaky key manager rotation failure test
In the key manager rotation failure test, start key managers after they
deregister to ensure that both are added to the committee simultaneously.
If key managers are started randomly, there is a risk that only one will
be added to the committee, as the other may register in another epoch.