go/storage/mkvs: Add PathBadger storage backend
Instead of using trie node hashes as keys in the underlying Badger
store, this new backend instead uses a combination of version and index
within the batch of trie nodes as keys which leads to improved locality
when iterating over the trie while at the same time making the database
smaller and compactions faster.

The new backend makes some (reasonable) assumptions, specifically that
only one root per type may be finalized in any version and that there
may be no child roots within the same version.

The new backend is experimental.