go/consensus/cometbft/light/service: Remove sparse light block store
If the CometBFT light client fails to find a light block at a specific
height, it will not attempt to fetch a neighboring one. Since the service
stores only one block per epoch, the likelihood of the light client
requesting that exact height is very low. This makes the trust store
from the service ineffective, especially considering that blocks are
no longer trusted once the trust period expires.