go/storage/mkvs: Stop extending path in doNext function
Extending path had no effect since old bitDepth is used,
which is correct given that this is for parent pointer.

Finally, to be consistent with other parts, recursive doNext
for the leaf node now passes newPath instead of just path.
Observe that bitDepth and path are actually ignored if passed
for the leaf node...