go/worker/storage/statesync: Pass context explicitly
In addition state sync worker should return an error and it should
be the caller responsibility to act accordingly. See e.g. new
workers such as stateless client.

Note that semantic changed slightly: Previously storage worker
would wait for all state sync workers to finish. Now it will
terminate when the first one finishes. Notice that this is not
100% true as previously state sync worker could panic (which
would in that case shutdown the whole node).