go/worker/storage/statesync: Refactor the code
The code was broken into smaller functions. Also the
scope of variables (including channels) have been reduced.

Semantics as well as performance should stay the same.

This refactor is needed if we want to make further changes
to the state. Ideally, the remaining logic in the Run method
should follow similar direction and user helpers to do db
initialization, checkpoint sync etc.