go/worker/storage/statesync: Remove redundant waitgroup
In case of termination due to error exiting main for loop or
canceled context there is no point in waiting for go routines
to finish fetching/doing the cleanup. As long we cancel the
context for them and use it properly in the select statements
this should be safe and better.