go/worker/storage: Move checkpointer to a separate worker.
Generally this worker should be tested explicitly.

For this checkpointer worker should probably take finalizedCh
instead of the whole state sync worker, waiting for the storage
initialization should be moved to the parent worker. Finally,
instead of the full common worker we should accept interface
to ease mocking.

Technically this worker does two things, creates and notifies.
I am open to moving checkpointer creation to the parent worker,
and notify there in the separate goroutine. This way this worker
is only responsible for creating checkpoints.

I have kept it this way as this way all the code relevant for
checkpoint creation or checkpointer functionality is encapsulated
together.