go/oasis-node/cmd/storage: Add command for offline pruning
When enabling aggresive pruning on a previously synced node and
restarting it immediately, node may start lagging behind (minutes
to hours) and still believe its status is ready.

We should offer validators a maintenance command that can be called
offline, when increasing or possibly enabling the pruning for the
first time, to ensure only healthy nodes join the network.