go/worker/storage: Move pruning out of the prune handler
By moving the pruning of the state db out of the prune handler,
we solve issue of nesting NodeDB.Prune (possibly long operation),
in the BadgerDB transaction created during the pruning of the
light history.

Futhermore, this also ensures that light history is always pruned
before the actual state. This is desired as previously light
history pruning had a side effect of deleting the state. If during
pruning there was an error the runtime history would get reverted,
but the state wouldn't. Looking at the `history.History`
`resolveRound`, this unlikely scenario was not anticipated, which
could result in `GetAnnotatedBlock` breaking its contract.

There is additional change of semantics: Previously if your state
had versions older than the earliest version in the light history
pruning would be stuck, as it would try to prune non-earliest
version from the NodeDB. Now, this is fixed but as a cosequence
this could silently start a very hot loop of pruning many versions.