go: Rename history.PruneHandler.Prune to CanPruneRuntime
This was done to be consistent with consensus prune handlers
terminology. Notice that consensus accepts a single round whereas
here we take a slice of rounds.

In my opinion pruning of light history in batch was a premature
optimization as time to prune light block << time to prune NodeDB
version.

For the follow-up we may want to unify this, possibly prune
handlers can be only checked for every n-th versions instead
of checking every versions (assuming that if pruning is valid
for N, than it should be also valid for N-1).