go/common/sgx/quote: Add new apply default policy method
Observe that already prior to this commit consensus and
runtime quote verification with regards to runtime constaints
were not symetric. Consensus part in addition to the runtime
constrains may also apply default constaints from the consensus
parameters. The runtime part on the other hand only applies
runtime constraints. Currently, this is not problematic as
as default policy as per the consensus parameters is nil,
but this may change in the future.

Nit: I would change to sc.ApplyDefaultConstraints(cfg.SGX),
as normaly the mutated part should be pointer receiver.