go/worker/compute/executor: Remove epoch check
The epoch check is incorrect because the group’s epoch
denotes the epoch in which the committee is valid, while
the block epoch represents the epoch in which the consensus
block was produced. When scheduling the first block of a new
epoch, these two epochs should differ by one.

Additionally, the check was optional because if the worker is
behind, a new block will cancel the ongoing round, resulting in
the same effect as the check itself.